pyproject.toml
[project]
name = "assist-bot"
version = "0.1.0"
description = "Alert Support & Solution Information in Slack Threads"
authors = [
    {name = "Diogo Regateiro",email = "diogo.regateiro@atlar.pt"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slackclient (>=2.9.4,<3.0.0)"
]

[tool.poetry.scripts]
assist = "assist_bot.bot:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
